wire up create blog to database
modify the routes that the blogs repository uses to have trailing
slashes.  That is needed for some reason.  Add a container around
all the blogs and give it space on the bottom so that there is some
space there when scrolled all the way down.  Have the controller
parse the result of each add and include it in the displayed blogs.